add clearScriptsCache function
