+++ shared-actions.js
@@ -10,5 +10,4 @@ module.exports.TIMER_END = 'logs/TIMER_END'
 module.exports.ADD_REQUEST = 'network/ADD_REQUEST'
 module.exports.SET_RESPONSE = 'network/SET_RESPONSE'
 module.exports.ADD_ACTION = 'actions/ADD_ACTION'
-module.exports.RUN_SCRIPT = 'playground/RUN_SCRIPT'
 module.exports.SET_SCRIPT_RESULT = 'playground/SET_SCRIPT_RESULT'
